Fix warnings

+++ drivers/video/msm/mdp.c
@@ -2547,8 +2547,8 @@ int mdp_bus_scale_update_request(u64 ab, u64 ib)
 
 	pr_info("%s: handle=%d index=%d ab=%d ib=%d\n", __func__,
 		 (u32)mdp_bus_scale_handle, bus_index,
-		 mdp_bus_usecases[bus_index].vectors->ab,
-		 mdp_bus_usecases[bus_index].vectors->ib);
+		 (int)mdp_bus_usecases[bus_index].vectors->ab,
+		 (int)mdp_bus_usecases[bus_index].vectors->ib);
 
 	return msm_bus_scale_client_update_request
 		(mdp_bus_scale_handle, bus_index);
